fix accessing rules

+++ src/tasks/tasks.service.ts
@@ -14,27 +14,27 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetFilterTaskDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(user: User, filterDto: GetFilterTaskDto): Promise<Task[]> {
+    return this.taskRepository.getTasks(user, filterDto);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(user: User, id: number): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
     if (!found) {
       throw new NotFoundException(`Task with ID: ${id} not found`);
     }
     return found;
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(user: User, id: number): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID: ${id} not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(user: User, id: number, status: TaskStatus): Promise<Task> {
+    const task = await this.getTaskById(user, id);
     task.status = status;
     await task.save();
     return task;
